created new layout for main page and pages to route the rest. At the moment this solves the conditional rendering of the NavBar and Footer.  Needs work
