+++ app/good/layout.js
@@ -1,0 +1,23 @@
+import { Inter } from 'next/font/google'
+import '../globals.css'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export const metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+}
+
+export default function Layout({ children, showNavBar = true }) {
+  return (
+    <html>
+      <body className={inter.className}>
+        <div className='content'>
+          <div className=''>
+            {children}
+          </div>
+        </div>
+      </body>
+    </html>
+  )
+}
