+++ app/pages/layout.js
@@ -1,7 +1,7 @@
 import { Navbar } from './Navbar.js'
 import Footer from './footer.js'
 import { Inter } from 'next/font/google'
-import './globals.css'
+import '../globals.css'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -10,16 +10,16 @@ export const metadata = {
   description: 'Generated by create next app',
 }
 
-export default function RootLayout({ children }) {
+export default function RootLayout({ children, showNavBar = true }) {
   return (
     <html>
       <body className={inter.className}>
         <div className='content'>
-          <Navbar />
+          {showNavBar && <Navbar />}
           <div className=''>
             {children}
           </div>
-          <Footer />
+          {showNavBar && <Footer />}
         </div>
       </body>
     </html>
