+++ app/pages/page.js
@@ -6,15 +6,14 @@ import Portfolio from './Portfolio.js'
 
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-2">
-      <div className='px-8 py-4'>
-        <Howdie />
-        <div className={`border-width: 2px rounded-lg border-color: rgb(255 255 255);`}>
-          <AboutMe />
-        </div>
-        <Services />
-        <Portfolio />
+    <main className="flex min-h-screen flex-col items-center justify-between p-2">      <div className='px-8 py-4'>
+      <Howdie />
+      <div className={`border-width: 2px rounded-lg border-color: rgb(255 255 255);`}>
+        <AboutMe />
       </div>
+      <Services />
+      <Portfolio />
+    </div>
       <div className="z-10 w-full max-w-5xl items-center justify-between font-mono text-sm lg:flex">
         {/* <p className="fixed left-0 top-0 flex w-full justify-center border-b border-gray-300 bg-gradient-to-b from-zinc-200 pb-6 pt-8 backdrop-blur-2xl dark:border-neutral-800 dark:bg-zinc-800/30 dark:from-inherit lg:static lg:w-auto  lg:rounded-xl lg:border lg:bg-gray-200 lg:p-4 lg:dark:bg-zinc-800/30">
           Get started by editing&nbsp;
@@ -122,6 +121,5 @@ export default function Home() {
       <div>
         {/* <Dashboard /> */}
       </div>
-    </main>
-  )
+    </main>)
 }
